making updates to the navigation demos so they are setup for vector.
Removed 3D for now until I have time to set it up. It was not using the most current 3D packages and is more effort to update

Added playpen example for sim

Updated vector_viz configurations

Updated laser naming to prepare for multiple laser support

Fixed issue with interactive marker and simulation not having a timeout when the user stops interacting with the teleop marker. Robot now stops after a 0.4 sec timeout.