updating the application
ESTOP now kills power but not the embedded system. When the ESTOP request is removed it will re-initialize the propulsion system for normal operation.

This does not tie into the ROS system, you will need to restart the service or make a monitoring node that monitors the topic

/vector/feedback/status

In particular the dynamic_response variable. When it is equal to 0x20 the system is in estop_mode, when it is 0 the system is operating normally with no detected faults that issue dynamic responses.

The monitoring node should detect this dynamic_response and reset the arm and gripper nodes when the E-Stop is cleared.