improved costmap inflation and planning behavior
planning was too strict on following global plan.
increasing the cost of not going to the goal makes
it easier to plan around unexpected obstacles.

inflation for global planner increased a lot becase
there is no reason to drive so close to everything.
the local planner is kept low because if you raise it,
then it gets very afraid of rotating in tight spaces,
which we need it to do.