adding the experimental gain tuning to a new branch until it is determined if any more updates to the behavior are needed
you need a key to unlock the ability to tune the control loops as it is dangerous. Any updates after it is unlocked are stored in NVM and persist between power cycles.

By default they are set middle of the road

The gains are set via the dynamic reconfigure server (such as using rqt_reconfigure or scripting it; if you want to change on the fly). For right now they are just updated via rqt_reconfigure.

When a new gain is set the machine will momentarily come to a stop and then update the gains; if unstable gains are accidentally sent, hit the E-STOP immediately and either reset to default gains or set the current ones to a stable state.

WARNING: if an unstable gainset is used it could potentially damage something or cause the robot to go out of control, care must be taken and liablilty is on the user to ensure you know what you are doing before adjusting gains. Always have your hand on the ESTOP when setting a new gain until you know it is stable.